Lifting configData up, and using as title
